Allow specifying desired data to fetch
By default, Google Scholar author search results return basic researcher
information, in chunks of 10 items per request. This data does not
include h-index. To obtain it, another request needs to be issued for
each author, which is a significant additional overhead.
Since h-index is not strictly necessary for author identification, we
don't need to fetch it initially. It is definitely needed later to
compile author statistics, so we need a way to request it. Furthermore,
we might need some other data (like coauthors, or recent/popular papers)
to aid the identification process later on.

To serve all of these needs, we add a parameter specifying the desired
information to the search function.