Modify the command registration mechanism
After this change, following a suggestion of argparse documentation we
store the command function itself in the parsing results, by using the
set_defaults function on each command parser. This makes "command
module" interface simpler, since it is now restricted to the
register_parser method. It also makes it possible to register multiple
commands with multiple handlers in one module, if necessary.